refactor: reorganize project structure for better maintainability
## 🏗️ Project Structure Reorganization

### Components Structure
- **ui/**: Reusable UI components (form fields, etc.)
  - `ui/form/` → Input.vue, Textarea.vue, Select.vue, Password.vue
- **features/**: Business logic components by domain
  - `features/post/` → Card.vue, CreateModal.vue, DeleteModal.vue
- **layout/**: Layout-specific components
  - LanguageSwitcher.vue, ThemeSwitcher.vue, PreferencesControls.vue

### Composables Structure
- **forms/**: Form-related composables
- **stores/**: Store wrapper composables
- **features/**: Feature-specific composables (notifications, SEO)

### Component Naming
Updated all component references to use Nuxt's auto-import naming:
- `<PostCard>` → `<FeaturesPostCard>`
- `<LanguageSwitcher>` → `<LayoutLanguageSwitcher>`
- `<FormFieldInput>` → `<UiFormInput>`
- etc.

### Bug Fixes
- ✅ Fixed SSR error `$setup.t is not a function` by moving useSeo() after useI18n()
- ✅ Recreated missing usePreferencesStore with proper Pinia setup
- ✅ Consolidated Post model with validation schemas in single file
- ✅ Updated all imports and component references

### Benefits
- 🎯 Clearer separation of concerns (UI vs Business Logic vs Layout)
- 🔧 Better maintainability with logical grouping
- 🚀 Consistent Nuxt auto-import naming convention
- 📦 Centralized Post model eliminates duplication
