+++ app/components/features/post/DeleteModal.vue
