+++ app/components/layout/PreferencesControls.vue
