+++ server/utils/validation.ts
@@ -1,5 +1,19 @@
-import type { ZodSchema } from 'zod'
+import type { ZodType, ZodError } from 'zod'
 import type { H3Event, EventHandlerRequest } from 'h3'
+
+/**
+ * Create a validation error from Zod result
+ */
+function createValidationError(message: string, error: ZodError) {
+  throw createError({
+    statusCode: 400,
+    statusMessage: 'Validation Failed',
+    data: {
+      message,
+      errors: error.issues
+    }
+  })
+}
 
 /**
  * Validate the request body against a Zod schema
@@ -11,26 +25,17 @@ import type { H3Event, EventHandlerRequest } from 'h3'
  */
 export async function validateBody<T>(
   event: H3Event<EventHandlerRequest>,
-  schema: ZodSchema<T>
+  schema: ZodType<T>
 ): Promise<T> {
   try {
     const body = await readBody(event)
     const result = schema.safeParse(body)
 
     if (!result.success) {
-      const formattedErrors = result.error.flatten()
-      throw createError({
-        statusCode: 400,
-        statusMessage: 'Validation Failed',
-        data: {
-          message: 'Validation failed',
-          fieldErrors: formattedErrors.fieldErrors,
-          formErrors: formattedErrors.formErrors
-        }
-      })
+      createValidationError('Request body validation failed', result.error)
     }
 
-    return result.data
+    return result.data!
   } catch (error: any) {
     if (error.statusCode) throw error
     throw createError({
@@ -50,26 +55,17 @@ export async function validateBody<T>(
  */
 export async function validateParams<T>(
   event: H3Event<EventHandlerRequest>,
-  schema: ZodSchema<T>
+  schema: ZodType<T>
 ): Promise<T> {
   try {
     const params = getRouterParams(event)
     const result = schema.safeParse(params)
 
     if (!result.success) {
-      const formattedErrors = result.error.flatten()
-      throw createError({
-        statusCode: 400,
-        statusMessage: 'Invalid Parameters',
-        data: {
-          message: 'Parameter validation failed',
-          fieldErrors: formattedErrors.fieldErrors,
-          formErrors: formattedErrors.formErrors
-        }
-      })
+      createValidationError('Parameter validation failed', result.error)
     }
 
-    return result.data
+    return result.data!
   } catch (error: any) {
     if (error.statusCode) throw error
     throw createError({
@@ -89,26 +85,17 @@ export async function validateParams<T>(
  */
 export async function validateQuery<T>(
   event: H3Event<EventHandlerRequest>,
-  schema: ZodSchema<T>
+  schema: ZodType<T>
 ): Promise<T> {
   try {
     const query = getQuery(event)
     const result = schema.safeParse(query)
 
     if (!result.success) {
-      const formattedErrors = result.error.flatten()
-      throw createError({
-        statusCode: 400,
-        statusMessage: 'Invalid Query',
-        data: {
-          message: 'Query validation failed',
-          fieldErrors: formattedErrors.fieldErrors,
-          formErrors: formattedErrors.formErrors
-        }
-      })
+      createValidationError('Query validation failed', result.error)
     }
 
-    return result.data
+    return result.data!
   } catch (error: any) {
     if (error.statusCode) throw error
     throw createError({
